+++ packages/api/src/Domain/Http/HttpHeaders.ts
@@ -1,1 +1,0 @@
-export type HttpHeaders = Map<string, string | null>