+++ packages/api/src/Domain/Http/HttpVerb.ts
@@ -1,7 +1,0 @@
-export enum HttpVerb {
-  Get = 'GET',
-  Post = 'POST',
-  Put = 'PUT',
-  Patch = 'PATCH',
-  Delete = 'DELETE',
-}