+++ packages/api/src/Domain/Server/Subscription/SubscriptionServer.ts
@@ -1,4 +1,3 @@
-import { AppleIAPConfirmResponse } from './../../Response/Subscription/AppleIAPConfirmResponse'
 import { HttpServiceInterface } from '../../Http/HttpServiceInterface'
 import { AppleIAPConfirmRequestParams } from '../../Request'
 import { SubscriptionInviteAcceptRequestParams } from '../../Request/Subscription/SubscriptionInviteAcceptRequestParams'
@@ -6,11 +5,13 @@ import { SubscriptionInviteCancelRequestParams } from '../../Request/Subscriptio
 import { SubscriptionInviteDeclineRequestParams } from '../../Request/Subscription/SubscriptionInviteDeclineRequestParams'
 import { SubscriptionInviteListRequestParams } from '../../Request/Subscription/SubscriptionInviteListRequestParams'
 import { SubscriptionInviteRequestParams } from '../../Request/Subscription/SubscriptionInviteRequestParams'
-import { SubscriptionInviteAcceptResponse } from '../../Response/Subscription/SubscriptionInviteAcceptResponse'
-import { SubscriptionInviteCancelResponse } from '../../Response/Subscription/SubscriptionInviteCancelResponse'
-import { SubscriptionInviteDeclineResponse } from '../../Response/Subscription/SubscriptionInviteDeclineResponse'
-import { SubscriptionInviteListResponse } from '../../Response/Subscription/SubscriptionInviteListResponse'
-import { SubscriptionInviteResponse } from '../../Response/Subscription/SubscriptionInviteResponse'
+import { AppleIAPConfirmResponseBody } from './../../Response/Subscription/AppleIAPConfirmResponseBody'
+import { SubscriptionInviteAcceptResponseBody } from '../../Response/Subscription/SubscriptionInviteAcceptResponseBody'
+import { SubscriptionInviteCancelResponseBody } from '../../Response/Subscription/SubscriptionInviteCancelResponseBody'
+import { SubscriptionInviteDeclineResponseBody } from '../../Response/Subscription/SubscriptionInviteDeclineResponseBody'
+import { SubscriptionInviteListResponseBody } from '../../Response/Subscription/SubscriptionInviteListResponseBody'
+import { SubscriptionInviteResponseBody } from '../../Response/Subscription/SubscriptionInviteResponseBody'
+import { HttpResponse } from '@standardnotes/responses'
 
 import { Paths } from './Paths'
 import { SubscriptionServerInterface } from './SubscriptionServerInterface'
@@ -18,39 +19,35 @@ import { SubscriptionServerInterface } from './SubscriptionServerInterface'
 export class SubscriptionServer implements SubscriptionServerInterface {
   constructor(private httpService: HttpServiceInterface) {}
 
-  async acceptInvite(params: SubscriptionInviteAcceptRequestParams): Promise<SubscriptionInviteAcceptResponse> {
-    const response = await this.httpService.post(Paths.v1.acceptInvite(params.inviteUuid), params)
-
-    return response as SubscriptionInviteAcceptResponse
+  async acceptInvite(
+    params: SubscriptionInviteAcceptRequestParams,
+  ): Promise<HttpResponse<SubscriptionInviteAcceptResponseBody>> {
+    return this.httpService.post(Paths.v1.acceptInvite(params.inviteUuid), params)
   }
 
-  async declineInvite(params: SubscriptionInviteDeclineRequestParams): Promise<SubscriptionInviteDeclineResponse> {
-    const response = await this.httpService.get(Paths.v1.declineInvite(params.inviteUuid), params)
-
-    return response as SubscriptionInviteDeclineResponse
+  async declineInvite(
+    params: SubscriptionInviteDeclineRequestParams,
+  ): Promise<HttpResponse<SubscriptionInviteDeclineResponseBody>> {
+    return this.httpService.get(Paths.v1.declineInvite(params.inviteUuid), params)
   }
 
-  async cancelInvite(params: SubscriptionInviteCancelRequestParams): Promise<SubscriptionInviteCancelResponse> {
-    const response = await this.httpService.delete(Paths.v1.cancelInvite(params.inviteUuid), params)
-
-    return response as SubscriptionInviteCancelResponse
+  async cancelInvite(
+    params: SubscriptionInviteCancelRequestParams,
+  ): Promise<HttpResponse<SubscriptionInviteCancelResponseBody>> {
+    return this.httpService.delete(Paths.v1.cancelInvite(params.inviteUuid), params)
   }
 
-  async listInvites(params: SubscriptionInviteListRequestParams): Promise<SubscriptionInviteListResponse> {
-    const response = await this.httpService.get(Paths.v1.listInvites, params)
-
-    return response as SubscriptionInviteListResponse
+  async listInvites(
+    params: SubscriptionInviteListRequestParams,
+  ): Promise<HttpResponse<SubscriptionInviteListResponseBody>> {
+    return this.httpService.get(Paths.v1.listInvites, params)
   }
 
-  async invite(params: SubscriptionInviteRequestParams): Promise<SubscriptionInviteResponse> {
-    const response = await this.httpService.post(Paths.v1.invite, params)
-
-    return response as SubscriptionInviteResponse
+  async invite(params: SubscriptionInviteRequestParams): Promise<HttpResponse<SubscriptionInviteResponseBody>> {
+    return this.httpService.post(Paths.v1.invite, params)
   }
 
-  async confirmAppleIAP(params: AppleIAPConfirmRequestParams): Promise<AppleIAPConfirmResponse> {
-    const response = await this.httpService.post(Paths.v1.confirmAppleIAP, params)
-
-    return response as AppleIAPConfirmResponse
+  async confirmAppleIAP(params: AppleIAPConfirmRequestParams): Promise<HttpResponse<AppleIAPConfirmResponseBody>> {
+    return this.httpService.post(Paths.v1.confirmAppleIAP, params)
   }
 }