+++ packages/api/src/Domain/Server/UserRequest/UserRequestServer.ts
@@ -1,6 +1,7 @@
 import { HttpServiceInterface } from '../../Http/HttpServiceInterface'
 import { UserRequestRequestParams } from '../../Request/UserRequest/UserRequestRequestParams'
-import { UserRequestResponse } from '../../Response/UserRequest/UserRequestResponse'
+import { HttpResponse } from '@standardnotes/responses'
+import { UserRequestResponseBody } from '../../Response/UserRequest/UserRequestResponseBody'
 
 import { Paths } from './Paths'
 import { UserRequestServerInterface } from './UserRequestServerInterface'
@@ -8,9 +9,7 @@ import { UserRequestServerInterface } from './UserRequestServerInterface'
 export class UserRequestServer implements UserRequestServerInterface {
   constructor(private httpService: HttpServiceInterface) {}
 
-  async submitUserRequest(params: UserRequestRequestParams): Promise<UserRequestResponse> {
-    const response = await this.httpService.post(Paths.v1.submitUserRequest(params.userUuid), params)
-
-    return response as UserRequestResponse
+  async submitUserRequest(params: UserRequestRequestParams): Promise<HttpResponse<UserRequestResponseBody>> {
+    return this.httpService.post(Paths.v1.submitUserRequest(params.userUuid), params)
   }
 }