+++ packages/mobile/android/app/src/main/java/com/standardnotes/MainApplication.java
@@ -2,12 +2,10 @@ package com.standardnotes;
 
 import android.app.Application;
 import android.app.Activity;
-import android.content.Context;
 import android.webkit.WebView;
 
 import com.facebook.react.PackageList;
 import com.facebook.react.ReactApplication;
-import com.facebook.react.ReactInstanceManager;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.defaults.DefaultNewArchitectureEntryPoint;
@@ -15,7 +13,6 @@ import com.facebook.react.defaults.DefaultReactNativeHost;
 import com.facebook.soloader.SoLoader;
 
 import java.lang.reflect.Field;
-import java.lang.reflect.InvocationTargetException;
 import java.util.List;
 
 import com.facebook.react.modules.network.OkHttpClientProvider;