+++ packages/responses/src/Domain/Http/HttpError.ts
@@ -1,0 +1,6 @@
+import { ErrorTag } from './ErrorTag'
+
+export type HttpError = {
+  message: string
+  tag?: ErrorTag
+}