+++ packages/responses/src/Domain/Http/HttpHeaders.ts
@@ -1,0 +1,1 @@
+export type HttpHeaders = Map<string, string | null>