+++ packages/responses/src/Domain/Http/HttpResponseMeta.ts
@@ -1,0 +1,11 @@
+import { Role } from '@standardnotes/security'
+
+export type HttpResponseMeta = {
+  auth: {
+    userUuid?: string
+    roles?: Role[]
+  }
+  server: {
+    filesServerUrl?: string
+  }
+}