+++ packages/responses/src/Domain/Item/GetSingleItemResponse.ts
@@ -1,14 +1,11 @@
-import { MinimalHttpResponse } from '../Http/MinimalHttpResponses'
 import { ServerItemResponse } from './ServerItemResponse'
 
-export type GetSingleItemResponse = MinimalHttpResponse & {
-  data:
-    | {
-        success: true
-        item: ServerItemResponse
-      }
-    | {
-        success: false
-        message: string
-      }
-}
+export type GetSingleItemResponse =
+  | {
+      success: true
+      item: ServerItemResponse
+    }
+  | {
+      success: false
+      message: string
+    }