+++ packages/responses/src/Domain/User/PostSubscriptionTokensResponse.ts
@@ -1,7 +1,3 @@
-import { MinimalHttpResponse } from '../Http/MinimalHttpResponses'
-
-export type PostSubscriptionTokensResponse = MinimalHttpResponse & {
-  data?: {
-    token: string
-  }
+export type PostSubscriptionTokensResponse = {
+  token: string
 }