+++ packages/desktop/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.105.10](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.150.9...@standardnotes/desktop@3.105.10) (2023-03-02)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.105.9](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.150.8...@standardnotes/desktop@3.105.9) (2023-03-02)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.105.8](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.150.7...@standardnotes/desktop@3.105.8) (2023-03-01)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.105.7](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.150.6...@standardnotes/desktop@3.105.7) (2023-03-01)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.105.6](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.150.5...@standardnotes/desktop@3.105.6) (2023-03-01)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.105.5](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.150.4...@standardnotes/desktop@3.105.5) (2023-03-01)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.105.4](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.150.3...@standardnotes/desktop@3.105.4) (2023-03-01)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
 ## [3.105.3](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.150.2...@standardnotes/desktop@3.105.3) (2023-02-27)
 
 **Note:** Version bump only for package @standardnotes/desktop