+++ packages/mobile/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.51.41](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.51.40...@standardnotes/mobile@3.51.41) (2023-03-02)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.51.40](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.51.39...@standardnotes/mobile@3.51.40) (2023-03-02)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.51.39](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.51.38...@standardnotes/mobile@3.51.39) (2023-03-01)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.51.38](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.51.37...@standardnotes/mobile@3.51.38) (2023-03-01)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.51.37](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.51.36...@standardnotes/mobile@3.51.37) (2023-03-01)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.51.36](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.51.35...@standardnotes/mobile@3.51.36) (2023-03-01)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.51.35](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.51.34...@standardnotes/mobile@3.51.35) (2023-03-01)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
 ## [3.51.34](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.51.33...@standardnotes/mobile@3.51.34) (2023-02-27)
 
 **Note:** Version bump only for package @standardnotes/mobile