+++ packages/web/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.150.10](https://github.com/standardnotes/app/compare/@standardnotes/web@3.150.9...@standardnotes/web@3.150.10) (2023-03-02)
+
+**Note:** Version bump only for package @standardnotes/web
+
+## [3.150.9](https://github.com/standardnotes/app/compare/@standardnotes/web@3.150.8...@standardnotes/web@3.150.9) (2023-03-02)
+
+**Note:** Version bump only for package @standardnotes/web
+
+## [3.150.8](https://github.com/standardnotes/app/compare/@standardnotes/web@3.150.7...@standardnotes/web@3.150.8) (2023-03-01)
+
+**Note:** Version bump only for package @standardnotes/web
+
+## [3.150.7](https://github.com/standardnotes/app/compare/@standardnotes/web@3.150.6...@standardnotes/web@3.150.7) (2023-03-01)
+
+**Note:** Version bump only for package @standardnotes/web
+
+## [3.150.6](https://github.com/standardnotes/app/compare/@standardnotes/web@3.150.5...@standardnotes/web@3.150.6) (2023-03-01)
+
+**Note:** Version bump only for package @standardnotes/web
+
+## [3.150.5](https://github.com/standardnotes/app/compare/@standardnotes/web@3.150.4...@standardnotes/web@3.150.5) (2023-03-01)
+
+**Note:** Version bump only for package @standardnotes/web
+
+## [3.150.4](https://github.com/standardnotes/app/compare/@standardnotes/web@3.150.3...@standardnotes/web@3.150.4) (2023-03-01)
+
+**Note:** Version bump only for package @standardnotes/web
+
 ## [3.150.3](https://github.com/standardnotes/app/compare/@standardnotes/web@3.150.2...@standardnotes/web@3.150.3) (2023-02-27)
 
 ### Bug Fixes